own author acl

+++ src/Application/Features/Articles/Commands/Delete.cs
@@ -3,6 +3,7 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
+using Application.Exceptions;
 using Application.Extensions;
 using Application.Features.Articles.Queries;
 using Application.Interfaces;
@@ -15,15 +16,22 @@ namespace Application.Features.Articles.Commands
     public class ArticleDeleteHandler : IAuthorizationRequestHandler<ArticleDeleteCommand>
     {
         private readonly IAppDbContext _context;
+        private readonly ICurrentUser _currentUser;
 
-        public ArticleDeleteHandler(IAppDbContext context)
+        public ArticleDeleteHandler(IAppDbContext context, ICurrentUser currentUser)
         {
             _context = context;
+            _currentUser = currentUser;
         }
 
         public async Task<Unit> Handle(ArticleDeleteCommand request, CancellationToken cancellationToken)
         {
             var article = await _context.Articles.FindAsync(x => x.Slug == request.Slug, cancellationToken);
+
+            if (article.AuthorId != _currentUser.User.Id)
+            {
+                throw new ForbiddenException();
+            }
 
             _context.Articles.Remove(article);
             await _context.SaveChangesAsync(cancellationToken);
